worker_v2/api_receive/pyproject.toml
[tool.poetry]
name = "api-receive"
version = "0.1.0"
description = ""
authors = ["Захаров Павел <pzakharov83@gmail.com>"]
readme = "README.md"
packages = [{include = "api_receive"}]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.1"
jinja2 = "^3.1.2"
python-barcode = "^0.14.0"
qrcode = "^7.4.2"
sqlalchemy = "^2.0.2"
pyodbc = "^4.0.35"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
pydantic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
